requirements.txt
requests
alpaca-py
google-api-python-client
google-auth
